Non-Functioning Anonymous Checkbox
This commit was entirely front end. The anonymous checkbox now appears in the toolbar
of the composer container (where you write posts). You can switch states of the box.

It's unclear to me still why the posts aren't appearing anonymous.
When the anonymous box is ticked and a post is submitted, it still shows the name and icon.
These could be either back-end or front-end issues. These will have to be tested.

1. I made a function to actually add in the checkbox to the toolbar (addAnonToggle)
2. I added "hooks" which are events driven situations that trigger the addition of this checkbox.
The hooks set off to add the checkbox when the composer container is opened, the page is
refreshed, and when a post is opened to be edited.

file changed:
public/src/client/compose.js